// codebase/frontend/app_/src/models/data/ILoginInfo.tsx
export interface ILoginInfo {
    address1: string;
    address2: string;
    birthDate: string;
    city: string;
    countryCode: string;
    dialingCode: string;
    easyJetMailingsFlag: boolean;
    email: string;
    firstName: string;
    lastName: string;
    mailingsFlag: boolean;
    mobilePhone: string;
    postalCode: string;
    title: string;
    preferredAirports?: string[];
}